@@ -3,6 +3,7 @@
 namespace App\Command;
 
 use App\Repository\Main\ApiKeyUserRepository;
+use Exception;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
@@ -15,8 +16,9 @@ use Symfony\Component\Console\Style\SymfonyStyle;
 )]
 class AuthListApikeyCommand extends Command
 {
-    public function __construct(private ApiKeyUserRepository $apiKeyUserRepository)
-    {
+    public function __construct(
+        private readonly ApiKeyUserRepository $apiKeyUserRepository
+    ) {
         parent::__construct();
     }
 
@@ -25,7 +27,7 @@ class AuthListApikeyCommand extends Command
     }
 
     /**
-     * @throws \Exception
+     * @throws Exception
      */
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
@@ -38,7 +40,7 @@ class AuthListApikeyCommand extends Command
             $id = $apiKey->getId();
             $name = $apiKey->getName();
 
-            $message[] = $id.'. '. $name;
+            $message[] = $id.'. '.$name;
         }
 
         $io->info(implode("\n", $message));
