@@ -3,6 +3,7 @@
 namespace App\Command;
 
 use App\Service\MicrosoftGraphServiceInterface;
+use Exception;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
@@ -15,15 +16,15 @@ use Symfony\Component\Console\Style\SymfonyStyle;
 )]
 class GraphCreateBookingCommand extends Command
 {
-    public function __construct(private MicrosoftGraphServiceInterface $microsoftGraphService)
-    {
+    public function __construct(
+        private readonly MicrosoftGraphServiceInterface $microsoftGraphService
+    ) {
         parent::__construct();
     }
 
-    protected function configure(): void
-    {
-    }
-
+    /**
+     * @throws Exception
+     */
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $io = new SymfonyStyle($input, $output);
