@@ -9,8 +9,9 @@ use App\Repository\Main\AAKResourceRepository;
 
 final class LocationCollectionDataProvider implements ContextAwareCollectionDataProviderInterface, RestrictedDataProviderInterface
 {
-    public function __construct(private AAKResourceRepository $AAKResourceRepository)
-    {
+    public function __construct(
+        private readonly AAKResourceRepository $AAKResourceRepository
+    ) {
     }
 
     public function supports(string $resourceClass, string $operationName = null, array $context = []): bool
