@@ -13,8 +13,11 @@ use Symfony\Component\Security\Core\Security;
 
 final class AAKResourceExtension implements QueryCollectionExtensionInterface, QueryItemExtensionInterface
 {
-    public function __construct(private Security $security, private RequestStack $requestStack, private CvrWhitelistRepository $cvrWhitelistRepository)
-    {
+    public function __construct(
+        private readonly Security $security,
+        private readonly RequestStack $requestStack,
+        private readonly CvrWhitelistRepository $cvrWhitelistRepository
+    ) {
     }
 
     public function applyToCollection(QueryBuilder $queryBuilder, QueryNameGeneratorInterface $queryNameGenerator, string $resourceClass, string $operationName = null)
